Rename `testing.utils` package to `testing.util`
