+++ subprojects/internal-testing/src/main/java/dev/nokee/internal/testing/util/ClosureTestUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.internal.testing.utils;
+package dev.nokee.internal.testing.util;
 
 import groovy.lang.Closure;
