+++ subprojects/internal-testing/src/test/java/dev/nokee/internal/testing/ConfigurationMatchers_ConsumableTest.java
@@ -19,7 +19,7 @@ import org.hamcrest.Matcher;
 import org.junit.jupiter.api.Test;
 
 import static dev.nokee.internal.testing.ConfigurationMatchers.consumable;
-import static dev.nokee.internal.testing.utils.ConfigurationTestUtils.testConfiguration;
+import static dev.nokee.internal.testing.util.ConfigurationTestUtils.testConfiguration;
 import static dev.nokee.utils.ConfigurationUtils.configureAsConsumable;
 
 class ConfigurationMatchers_ConsumableTest extends AbstractMatcherTest {
