+++ subprojects/internal-testing/src/test/java/dev/nokee/internal/testing/ConfigurationMatchers_DeclarableTest.java
@@ -19,7 +19,7 @@ import org.hamcrest.Matcher;
 import org.junit.jupiter.api.Test;
 
 import static dev.nokee.internal.testing.ConfigurationMatchers.declarable;
-import static dev.nokee.internal.testing.utils.ConfigurationTestUtils.testConfiguration;
+import static dev.nokee.internal.testing.util.ConfigurationTestUtils.testConfiguration;
 import static dev.nokee.utils.ConfigurationUtils.configureAsDeclarable;
 
 class ConfigurationMatchers_DeclarableTest extends AbstractMatcherTest {
