+++ subprojects/platform-base/src/test/groovy/dev/nokee/platform/base/internal/dependencies/ProjectConfigurationActions_ExtendsFromTest.java
@@ -21,7 +21,7 @@ import org.junit.jupiter.api.Test;
 import spock.lang.Subject;
 
 import static dev.nokee.internal.testing.GradleNamedMatchers.named;
-import static dev.nokee.internal.testing.utils.ConfigurationTestUtils.testConfiguration;
+import static dev.nokee.internal.testing.util.ConfigurationTestUtils.testConfiguration;
 import static dev.nokee.platform.base.internal.dependencies.ProjectConfigurationActions.assertConfigured;
 import static dev.nokee.platform.base.internal.dependencies.ProjectConfigurationActions.extendsFrom;
 import static org.hamcrest.MatcherAssert.assertThat;
